+++ src/com/landawn/abacus/util/Try.java
@@ -345,6 +345,11 @@ public final class Try<T extends AutoCloseable> {
         boolean test(A a, B b, C c) throws E;
     }
 
+    @SuppressWarnings("hiding")
+    public static interface QuadPredicate<A, B, C, D, E extends Exception> {
+        boolean test(A a, B b, C c, D d) throws E;
+    }
+
     public static interface Function<T, R, E extends Exception> {
         R apply(T t) throws E;
     }
@@ -357,6 +362,11 @@ public final class Try<T extends AutoCloseable> {
         R apply(A a, B b, C c) throws E;
     }
 
+    @SuppressWarnings("hiding")
+    public interface QuadFunction<A, B, C, D, R, E extends Exception> {
+        R apply(A a, B b, C c, D d) throws E;
+    }
+
     public static interface Consumer<T, E extends Exception> {
         void accept(T t) throws E;
     }
@@ -367,5 +377,10 @@ public final class Try<T extends AutoCloseable> {
 
     public static interface TriConsumer<A, B, C, E extends Exception> {
         void accept(A a, B b, C c) throws E;
+    }
+
+    @SuppressWarnings("hiding")
+    public interface QuadConsumer<A, B, C, D, E extends Exception> {
+        void accept(A a, B b, C c, D d) throws E;
     }
 }