+++ src/com/landawn/abacus/util/stream/AbstractByteStream.java
@@ -601,28 +601,12 @@ abstract class AbstractByteStream extends ByteStream {
 
     @Override
     public String join(final CharSequence delimiter, final CharSequence prefix, final CharSequence suffix) {
-        final Supplier<Joiner> supplier = new Supplier<Joiner>() {
-            @Override
-            public Joiner get() {
-                return Joiner.with(delimiter, prefix, suffix);
-            }
-        };
+        final Joiner joiner = Joiner.with(delimiter, prefix, suffix).reuseStringBuilder(true);
+        final ByteIteratorEx iter = this.iteratorEx();
 
-        final ObjByteConsumer<Joiner> accumulator = new ObjByteConsumer<Joiner>() {
-            @Override
-            public void accept(Joiner a, byte t) {
-                a.append(t);
-            }
-        };
-
-        final BiConsumer<Joiner, Joiner> combiner = new BiConsumer<Joiner, Joiner>() {
-            @Override
-            public void accept(Joiner a, Joiner b) {
-                a.merge(b);
-            }
-        };
-
-        final Joiner joiner = collect(supplier, accumulator, combiner);
+        while (iter.hasNext()) {
+            joiner.append(iter.nextByte());
+        }
 
         return joiner.toString();
     }