+++ src/com/landawn/abacus/util/stream/AbstractDoubleStream.java
@@ -663,28 +663,12 @@ abstract class AbstractDoubleStream extends DoubleStream {
 
     @Override
     public String join(final CharSequence delimiter, final CharSequence prefix, final CharSequence suffix) {
-        final Supplier<Joiner> supplier = new Supplier<Joiner>() {
-            @Override
-            public Joiner get() {
-                return Joiner.with(delimiter, prefix, suffix);
-            }
-        };
+        final Joiner joiner = Joiner.with(delimiter, prefix, suffix).reuseStringBuilder(true);
+        final DoubleIteratorEx iter = this.iteratorEx();
 
-        final ObjDoubleConsumer<Joiner> accumulator = new ObjDoubleConsumer<Joiner>() {
-            @Override
-            public void accept(Joiner a, double t) {
-                a.append(t);
-            }
-        };
-
-        final BiConsumer<Joiner, Joiner> combiner = new BiConsumer<Joiner, Joiner>() {
-            @Override
-            public void accept(Joiner a, Joiner b) {
-                a.merge(b);
-            }
-        };
-
-        final Joiner joiner = collect(supplier, accumulator, combiner);
+        while (iter.hasNext()) {
+            joiner.append(iter.nextDouble());
+        }
 
         return joiner.toString();
     }