+++ src/com/landawn/abacus/util/stream/AbstractLongStream.java
@@ -598,28 +598,12 @@ abstract class AbstractLongStream extends LongStream {
 
     @Override
     public String join(final CharSequence delimiter, final CharSequence prefix, final CharSequence suffix) {
-        final Supplier<Joiner> supplier = new Supplier<Joiner>() {
-            @Override
-            public Joiner get() {
-                return Joiner.with(delimiter, prefix, suffix);
-            }
-        };
+        final Joiner joiner = Joiner.with(delimiter, prefix, suffix).reuseStringBuilder(true);
+        final LongIteratorEx iter = this.iteratorEx();
 
-        final ObjLongConsumer<Joiner> accumulator = new ObjLongConsumer<Joiner>() {
-            @Override
-            public void accept(Joiner a, long t) {
-                a.append(t);
-            }
-        };
-
-        final BiConsumer<Joiner, Joiner> combiner = new BiConsumer<Joiner, Joiner>() {
-            @Override
-            public void accept(Joiner a, Joiner b) {
-                a.merge(b);
-            }
-        };
-
-        final Joiner joiner = collect(supplier, accumulator, combiner);
+        while (iter.hasNext()) {
+            joiner.append(iter.nextLong());
+        }
 
         return joiner.toString();
     }